refactor: extract stream utilities to resolve circular dependencies and inline requires
