+++ src/downloadUtils.ts
@@ -6,6 +6,7 @@ import * as tmp from "tmp";
 
 import { Client } from "./apiv2";
 import { FirebaseError } from "./error";
+import { streamToString } from "./streamUtils";
 
 /**
  * Downloads the resource at `remoteUrl` to a temporary file.
@@ -28,7 +29,6 @@ export async function downloadToTmp(remoteUrl: string, auth = false): Promise<st
     resolveOnHTTPError: true,
   });
   if (res.status !== 200) {
-    const { streamToString } = require("./utils");
     const errorText = await streamToString(res.body);
     throw new FirebaseError(`download failed, status ${res.status}: ${errorText}`, {
       status: res.status,
