+++ src/utils.ts
@@ -12,7 +12,6 @@ import * as clc from "colorette";
 import * as open from "open";
 import * as ora from "ora";
 import * as process from "process";
-import { Readable } from "stream";
 import { AssertionError } from "assert";
 import { getPortPromise as getPort } from "portfinder";
 
@@ -309,34 +308,7 @@ export function explainStdin(): void {
   }
 }
 
-/**
- * Converts text input to a Readable stream.
- * @param text string to turn into a stream.
- * @return Readable stream, or undefined if text is empty.
- */
-export function stringToStream(text: string): Readable | undefined {
-  if (!text) {
-    return undefined;
-  }
-  const s = new Readable();
-  s.push(text);
-  s.push(null);
-  return s;
-}
-
-/**
- * Converts a Readable stream into a string.
- * @param s a readable stream.
- * @return a promise resolving to the string'd contents of the stream.
- */
-export function streamToString(s: NodeJS.ReadableStream): Promise<string> {
-  return new Promise((resolve, reject) => {
-    let b = "";
-    s.on("error", reject);
-    s.on("data", (d) => (b += `${d}`));
-    s.once("end", () => resolve(b));
-  });
-}
+export { stringToStream, streamToString } from "./streamUtils";
 
 /**
  * Sets the active project alias or id in the specified directory.
